requirements.txt
fastapi
uvicorn[standard]
python-docx
pandas
openpyxl
requests
msal
python-multipart
docxtpl
Pillow